+++ app/repository/CarAdvertsRepository.scala
@@ -6,12 +6,12 @@ trait CarAdvertsRepository {
 
   def findAll(): List[CarAdvert]
 
-  def find(id: Long): Option[CarAdvert]
+  def find(id: Int): Option[CarAdvert]
 
   def create(carAdvert: CarAdvert): Boolean
 
-  def update(id: Long, carAdvert: CarAdvert): Boolean
+  def update(id: Int, carAdvert: CarAdvert): Boolean
 
-  def delete(id: Long): Unit
+  def delete(id: Int): Unit
 
 }
