+++ app/repository/MongoCarAdvertsRepository.scala
@@ -16,7 +16,7 @@ class MongoCarAdvertsRepository(db: DB = CarAdvertsDB) extends CarAdvertsReposit
   def findAll(): List[CarAdvert] = 
     db.carAdvertsCollection.find().flatMap(_.asCarAdvert).toList
 
-  def find(id: Long): Option[CarAdvert] = {
+  def find(id: Int): Option[CarAdvert] = {
     val query = queryById(id)
 
     db.carAdvertsCollection.findOne(query).flatMap(_.asCarAdvert)
@@ -25,19 +25,19 @@ class MongoCarAdvertsRepository(db: DB = CarAdvertsDB) extends CarAdvertsReposit
   def create(carAdvert: CarAdvert): Boolean = 
     db.carAdvertsCollection.insert(carAdvert.asDBObject, WriteConcern.Safe).wasAcknowledged()
 
-  def update(id: Long, carAdvert: CarAdvert): Boolean = {
+  def update(id: Int, carAdvert: CarAdvert): Boolean = {
     val query = queryById(id)
 
     db.carAdvertsCollection.update(query, carAdvert.asDBObject).wasAcknowledged()
   }
 
-  def delete(id: Long): Unit = {
+  def delete(id: Int): Unit = {
     val query = queryById(id)
 
     db.carAdvertsCollection.findAndRemove(query)
   }
 
-  private def queryById(id: Long) = {
+  private def queryById(id: Int) = {
     MongoDBObject("id" -> id)
   }
